Switch from password to key
On further consideration, it doesn't seem like there is any benefit to
using a human-generated password for `cryogen`'s encryption/signing
procedures. If a password is provided with enough entropy (512 bits in
total) it won't be easily memorisable anyway, and if a password is
provided with too little entropy then the complexity required to stretch
it and preserve the salt is not worth it. Given `cryogen`'s use case of
storing secrets securely in-repo, it seems reasonable to generate a
secure-by-default key and require the user to store that key somewhere
(password manager, etc.) instead. Switching to this approach simplifies
the overall approach quite a bit and ensures that lock/unlock cycles
work seamlessly right off the bat.